+++ project/src/binarization/integral_binarization.cpp
@@ -22,8 +22,8 @@
 
 void computeIntegralImages(const unsigned char* gray,
                            int width, int height,
-                           std::vector<double>& integralImg,
-                           std::vector<double>& integralImgSq)
+                           std::vector<float>& integralImg,
+                           std::vector<float>& integralImgSq)
 {
     integralImg.resize(width * height, 0.0);
     integralImgSq.resize(width * height, 0.0);
@@ -31,9 +31,9 @@ void computeIntegralImages(const unsigned char* gray,
     // 1. Row-wise scan (prefix sums per row)
 #pragma omp parallel for
     for (int y = 0; y < height; y++) {
-        double sumRow = 0.0, sumRowSq = 0.0;
+        float sumRow = 0.0, sumRowSq = 0.0;
         for (int x = 0; x < width; x++) {
-            double val = static_cast<double>(gray[y * width + x]);
+            float val = static_cast<float>(gray[y * width + x]);
             sumRow   += val;
             sumRowSq += val * val;
             integralImg[y * width + x]   = sumRow;
@@ -44,7 +44,7 @@ void computeIntegralImages(const unsigned char* gray,
     // 2. Column-wise scan (prefix sums of the already row-summed data)
 #pragma omp parallel for
     for (int x = 0; x < width; x++) {
-        double colSum = 0.0, colSumSq = 0.0;
+        float colSum = 0.0, colSumSq = 0.0;
         for (int y = 0; y < height; y++) {
             colSum   += integralImg[y * width + x];
             colSumSq += integralImgSq[y * width + x];
@@ -65,7 +65,7 @@ void computeIntegralImages(const unsigned char* gray,
  * @return Sum of pixel values in the region.
  */
 
-inline double getSum(const std::vector<double>& integralImg,
+inline float getSum(const std::vector<float>& integralImg,
                      int x1, int y1, int x2, int y2, int width, int height)
 {
     // Clamping region boundaries
@@ -74,10 +74,10 @@ inline double getSum(const std::vector<double>& integralImg,
     if (y2 >= height) y2 = height - 1;
 
     // Using the inclusion-exclusion principle to compute region sum efficiently
-    double A = (x1 > 0 && y1 > 0) ? integralImg[(y1 - 1) * width + (x1 - 1)] : 0.0;
-    double B = (y1 > 0) ? integralImg[(y1 - 1) * width + x2] : 0.0;
-    double C = (x1 > 0) ? integralImg[y2 * width + (x1 - 1)] : 0.0;
-    double D = integralImg[y2 * width + x2];
+    float A = (x1 > 0 && y1 > 0) ? integralImg[(y1 - 1) * width + (x1 - 1)] : 0.0;
+    float B = (y1 > 0) ? integralImg[(y1 - 1) * width + x2] : 0.0;
+    float C = (x1 > 0) ? integralImg[y2 * width + (x1 - 1)] : 0.0;
+    float D = integralImg[y2 * width + x2];
     return D + A - B - C;
 }
 
@@ -94,21 +94,21 @@ inline double getSum(const std::vector<double>& integralImg,
  * @param stddev Output standard deviation value.
  */
 
-void local_mean_std_integral(const std::vector<double>& integralImg,
-                             const std::vector<double>& integralImgSq,
+void local_mean_std_integral(const std::vector<float>& integralImg,
+                             const std::vector<float>& integralImgSq,
                              int width, int height,
                              int x, int y, int half_win,
                              float &mean, float &stddev)
 {
     int x1 = x - half_win, y1 = y - half_win;
     int x2 = x + half_win, y2 = y + half_win;
-    double area = (x2 - x1 + 1) * (y2 - y1 + 1);
+    float area = (x2 - x1 + 1) * (y2 - y1 + 1);
 
-    double sum = getSum(integralImg, x1, y1, x2, y2, width, height);
-    double sumSq = getSum(integralImgSq, x1, y1, x2, y2, width, height);
+    float sum = getSum(integralImg, x1, y1, x2, y2, width, height);
+    float sumSq = getSum(integralImgSq, x1, y1, x2, y2, width, height);
 
-    double m = sum / area;
-    double var = (sumSq / area) - (m * m);
+    float m = sum / area;
+    float var = (sumSq / area) - (m * m);
     mean = static_cast<float>(m);
     stddev = (var > 0.0) ? static_cast<float>(std::sqrt(var)) : 0.0f;
 }
@@ -121,8 +121,8 @@ void adaptive_binarize_integral(const unsigned char* gray,
                        unsigned char* out,
                        int width, int height,
                        int window_size,
-                       const std::vector<double>& integralImg,
-                       const std::vector<double>& integralImgSq,
+                       const std::vector<float>& integralImg,
+                       const std::vector<float>& integralImgSq,
                        const std::function<float(float mean, float stddev)> &threshold_func) {
     int half_win = window_size / 2;
 
@@ -141,7 +141,7 @@ void adaptive_binarize_integral(const unsigned char* gray,
     }
 
     auto end = std::chrono::high_resolution_clock::now();
-    std::chrono::duration<double> duration = end - start;
+    std::chrono::duration<float> duration = end - start;
     spdlog::info("Adaptive integral binarization completed in {} seconds.", duration.count());
 }
 
@@ -155,8 +155,8 @@ void sauvola_binarize_integral(const unsigned char* gray,
                                int window_size,
                                float k,
                                float R,
-                               const std::vector<double>& integralImg,
-                               const std::vector<double>& integralImgSq) {
+                               const std::vector<float>& integralImg,
+                               const std::vector<float>& integralImgSq) {
     spdlog::info("Starting Integral Sauvola binarization with window size {}, k={}, R={}.", window_size, k, R);
 
     auto threshold_func = [k, R](float mean, float stddev) {
@@ -190,7 +190,7 @@ void process_integral_binarization(const std::string &input_path, int window_siz
                 0.0722f * image[i * channels + 2]);
     }
 
-    std::vector<double> integralImg, integralImgSq;
+    std::vector<float> integralImg, integralImgSq;
     computeIntegralImages(gray.data(), width, height, integralImg, integralImgSq);
 
     std::string output_path_integral = make_output_path(input_path) + "_integral_sauvola.png";
@@ -208,7 +208,7 @@ void process_integral_binarization(const std::string &input_path, int window_siz
     }
 
     auto end = std::chrono::high_resolution_clock::now();
-    std::chrono::duration<double> duration = end - start;
+    std::chrono::duration<float> duration = end - start;
     spdlog::info("Integral binarization process completed in {} seconds.", duration.count());
 
     stbi_image_free(image);
